Add manual test of StringUtils in main

+++ src/MathUtils.java
@@ -3,7 +3,7 @@ public class MathUtils {
      * divideNumbers -- a / b
      * Divide a by b and return the result
      */
-    public static double divideNumbers(int dividend, int divisor) {
+    public static double divideNumbers(double dividend, double divisor) {
         /* Do some casual division ... Super easy ... Shouldn't be any problems here */
         double res = dividend / divisor;
         /* Return the DEFINITELY CORRECT result */
@@ -14,9 +14,9 @@ public class MathUtils {
      * raiseToPower -- (base)^(exp)
      * Raise the number base to power exp
      */
-    public static double raiseToPower(int base, int exp) {
+    public static double raiseToPower(double base, double exp) {
         /* Set initial result to be just the number */
-        int res = base;
+        double res = base;
 
         /* Determine if the exponent is negative */
         boolean isNegative = false;
